// src/census/enum/census-variable.enum.ts
export enum CensusVariable {
  TOTAL_POP = 'S0101_C01_001E',
  TOTAL_UNDER_18_POP = 'S0101_C01_022E',
  TOTAL_UNDER_18_MALE_POP = 'S0101_C03_022E',
  TOTAL_UNDER_18_FEMALE_POP = 'S0101_C05_022E',
  TOTAL_POVERTY_STATUS_UNDER_6 = 'B17020_003E',
  TOTAL_POVERTY_STATUS_6_TO_11 = 'B17020_004E',
  TOTAL_POVERTY_STATUS_12_TO_17 = 'B17020_005E',
}